+++ zoom/src/menu.c
@@ -42,7 +42,7 @@
 
 static void center(char* text, int columns)
 {
-  display_set_cursor((columns>>1)-((int) strlen(text)>>1), display_get_cur_y());
+  display_set_cursor((int)((columns>>1)-(strlen(text)>>1)), display_get_cur_y());
   display_prints_c(text);
 }
 
@@ -94,9 +94,9 @@ char* menu_get_story(void)
   /* Read the files in this directory, and work out their names */
   while ((dent=readdir(gamedir)))
     {
-      int len;
+      size_t len;
 
-      len = (int) strlen(dent->d_name);
+      len = strlen(dent->d_name);
       
       if (len > 2)
 	{
@@ -108,7 +108,7 @@ char* menu_get_story(void)
 	      
 	      if (file)
 		{
-		  int x,len;
+		  size_t x,len;
 		  ZByte* header;
 
 		  header = read_block(file, 0, 64);
@@ -125,7 +125,7 @@ char* menu_get_story(void)
 
 		  if (game[n_games].storyname == NULL)
 		    {
-		      len = (int) strlen(game[n_games].filename);
+		      len = strlen(game[n_games].filename);
 		      
 		      game[n_games].storyname = malloc(len+1);
 		      for (x=0; x<len-3; x++)
